requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
crispy-bootstrap5==2024.2
Django==5.0.6
django-crispy-forms==2.2
gunicorn==22.0.0
packaging==24.1
pillow==10.4.0
sqlparse==0.5.0
tzdata==2024.1